requirements.txt
flask
spacy
numpy
pymongo
gunicorn
python-dotenv
